# PERF: Categorical getitem perf