Convert to an array earlier on.
Closes https://github.com/pandas-dev/pandas/issues/30744