@TomAugspurger 2020-01-06
Master

```python
   ...: %timeit data[list_]
1.44 ms ± 5.3 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

```

This PR

```
   ...: %timeit data[list_]
740 µs ± 5.92 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

```

Will post the pre-indexing with NA timing in a bit.