@TomAugspurger 2020-01-06
> can you add an asv

I think Joris caught this because the existing asv `CategoricalSlicing.time_getitem_list` https://pandas.pydata.org/speed/pandas/#categoricals.CategoricalSlicing.time_getitem_list?commits=6efc2379-b9de33e3