@jorisvandenbossche 2020-01-06
So I reported this for Categorical (indeed because of the existing benchmark), but a similar change was made for other dtypes in https://github.com/pandas-dev/pandas/pull/30308. Also the boolean and integer (and numpy) arrays have the same issue.